feat (FORCE) change the way forceLink creates links
